requirements.txt
anyio==3.6.2
alembic==1.9.1
asyncpg==0.27.0
azure-cognitiveservices-vision-computervision==0.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
fastapi-utils==0.2.1
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.22.4
networkx==2.8.8
ninja==1.11.1
packaging==22.0
Pillow==9.3.0
pyclipper==1.3.0.post4
pydantic==1.10.2
python-bidi==0.4.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyWavelets==1.4.1
PyYAML==6.0
psycopg2-binary==2.9.5
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
SQLAlchemy==1.4.45
tifffile==2022.10.10
torch==1.13.0
torchvision==0.10.0
typing_extensions==4.4.0
unicorn==2.0.1.post1
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
